Improved Refactor to take other features into account
